pyproject.toml
[project]
name = "manual-render"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipython>=9.4.0",
    "moderngl>=5.12.0",
    "numpy>=2.3.2",
    "pygame>=2.6.1",
    "pyqt5>=5.15.11",
    "pysdl2>=0.9.17",
    "pysdl2-dll>=2.32.0",
    "watchdog>=6.0.0",
    "watchfiles>=1.1.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.17.1",
    "ruff>=0.12.7",
]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "C90", "I", "N", "UP", "B", "A", "C4", "T20"]
ignore = ["E501", "C901"]  # Line too long (handled by black), complexity check

[tool.black]
line-length = 88
target-version = ['py313']

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Start lenient for graphics code
